requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.11.8
    # via web3
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
async-timeout==5.0.1
    # via aiohttp
attrs==24.2.0
    # via aiohttp
bitarray==3.0.0
    # via eth-account
black==24.10.0
    # via -r requirements.in
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
ckzg==2.0.1
    # via eth-account
click==8.1.7
    # via
    #   -r requirements.in
    #   black
cytoolz==1.0.0
    # via eth-utils
eth-abi==5.1.0
    # via
    #   eth-account
    #   web3
eth-account==0.13.4
    # via web3
eth-hash[pycryptodome]==0.7.0
    # via
    #   eth-utils
    #   web3
eth-keyfile==0.8.1
    # via eth-account
eth-keys==0.6.0
    # via
    #   eth-account
    #   eth-keyfile
eth-rlp==2.1.0
    # via eth-account
eth-typing==5.0.1
    # via
    #   -r requirements.in
    #   eth-abi
    #   eth-keys
    #   eth-utils
    #   web3
eth-utils==5.1.0
    # via
    #   -r requirements.in
    #   eth-abi
    #   eth-account
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   rlp
    #   web3
exceptiongroup==1.2.2
    # via pytest
flake8==7.1.1
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
hexbytes==1.2.1
    # via
    #   eth-account
    #   eth-rlp
    #   web3
idna==3.10
    # via
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
mccabe==0.7.0
    # via flake8
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
packaging==24.2
    # via
    #   black
    #   pytest
parsimonious==0.10.0
    # via eth-abi
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via black
pluggy==1.5.0
    # via pytest
propcache==0.2.0
    # via
    #   aiohttp
    #   yarl
pycodestyle==2.12.1
    # via flake8
pycryptodome==3.21.0
    # via
    #   eth-hash
    #   eth-keyfile
pydantic==2.10.2
    # via
    #   eth-account
    #   web3
pydantic-core==2.27.1
    # via pydantic
pyflakes==3.2.0
    # via flake8
pytest==8.3.3
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
pyunormalize==16.0.0
    # via web3
regex==2024.11.6
    # via parsimonious
requests==2.32.3
    # via web3
rlp==4.0.1
    # via
    #   eth-account
    #   eth-rlp
tomli==2.2.1
    # via
    #   black
    #   pytest
toolz==1.0.0
    # via cytoolz
types-requests==2.32.0.20241016
    # via web3
typing-extensions==4.12.2
    # via
    #   black
    #   eth-rlp
    #   eth-typing
    #   multidict
    #   pydantic
    #   pydantic-core
    #   web3
urllib3==2.2.3
    # via
    #   requests
    #   types-requests
web3==7.6.0
    # via -r requirements.in
websockets==13.1
    # via web3
yarl==1.18.0
    # via aiohttp
